+++ packages/styled-components/mixin.core.js
@@ -5,9 +5,11 @@ class StyledComponentsMixin extends Mixin {
     const { experimentalEsbuild } = this.options;
 
     if (experimentalEsbuild) {
-      console.warn(
-        'The experimental esbuild transpilation does not support styled components yet!'
-      );
+      if (typeof this.getLogger === 'function') {
+        this.getLogger().warn(
+          'The experimental esbuild transpilation does not support styled components yet!'
+        );
+      }
       return;
     }